requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
fastapi==0.115.7
h11==0.14.0
idna==3.10
iniconfig==2.0.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
sniffio==1.3.1
starlette==0.45.2
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
